Huge refactoring
An Expression now gets bound to a particular data type and can only be run on that type. Validations of the match operations is done prior to executing the first match. This is either done with the CreateForType function or lazily when the first Evaluate invocation is made.

The evaluation code makes lots of assumptions that everything has been validated and I tried to document those assumptions.