Add the ability to specify collection values inline
This adds a new selector type that uses `[a, b]` syntax to specify
values in-line as a collection rather than via a path to an object. This
can be used for circumstances such as

`"/userinfo/email" in ["foo@example.com", "bar@example.com"]`

The values specified are StringLiterals so must be quoted but normal
coalescing rules should apply.